+++ ChatApp/templates/pages/register-pages/login.html
@@ -8,7 +8,7 @@
                 <h1>Welcome to <br>
                     ShareHappy</h1>
             <div class="registration-label">
-                <label for="username">ユーザーネーム</label>
+                <label for="username">メールアドレス</label>
                 <input type="text" id="username" name="username">
             </div>
             <div class="registration-label">
